Fix Load error due to nested selects.
A select that takes a subset of vector lanes from an existing
subset of vector lanes tries to use the mangled name of the
subselected vector (e.g. foo.rgb), leading to an error.
Instead, introduce a temporary and rely on the OpenGL compiler
to optimize this away via CSE.